fix(cuda): integer-type precision in picture_cuda.c (round-5 clang-tidy) (#704)
Fixes five `bugprone-narrowing-conversions` /
`bugprone-implicit-widening-of-multiplication-result` findings in
`libvmaf/src/cuda/picture_cuda.c`, surfaced by the round-5 clang-tidy
`bugprone-*` sweep on CUDA host TUs.

**Defects corrected**

1. `vmaf_cuda_picture_download_async` – `CUDA_MEMCPY2D.WidthInBytes`
   (`size_t`) computed from a bare `unsigned × unsigned` product; for
   frames wider than ~2 GiB the product wraps at 2³² before widening.
   Fixed: `(size_t)cuda_pic->w[i] * ((pic->bpc + 7) / 8)`.

2. `vmaf_cuda_picture_upload_async` – same issue. Fixed identically.

3. `vmaf_cuda_picture_alloc` – `cuMemAllocPitch` width argument had the
   same `unsigned × unsigned` overflow path. Fixed identically.

4. `vmaf_cuda_picture_alloc_pinned` – `aligned_y` and `aligned_c` were
   `int` but computed from `unsigned` arithmetic; the bitwise NOT mask
   `~(DATA_ALIGN_PINNED - 1)` evaluated as a signed complement, making
   the expression UB when the alignment arithmetic exceeded `INT_MAX`.
   Changed to `unsigned` with explicit `1u` mask literal.

5. `vmaf_cuda_picture_free` – `vmaf_ref_load()` returns `long` (64-bit
   atomic counter) but the result was narrowed to `int`. Changed to
   `long err`.

**Scope**: CUDA host TU only; no kernel (.cu) changes, no scoring
changes, no public API changes. No Netflix golden assertions touched.

**Reproducer**:
  clang-tidy \
    -checks='-*,bugprone-narrowing-conversions,bugprone-implicit-widening-of-multiplication-result' \
    -p libvmaf/build-cuda \
    libvmaf/src/cuda/picture_cuda.c
(zero warnings after this fix)

no rebase impact: CUDA host TU type fixes only; no upstream API surface

+++ libvmaf/src/cuda/picture_cuda.c
@@ -46,7 +46,7 @@ int vmaf_cuda_picture_download_async(VmafPicture *cuda_pic, VmafPicture *pic, ui
         m.srcPitch = cuda_pic->stride[i];
         m.dstHost = pic->data[i];
         m.dstPitch = pic->stride[i];
-        m.WidthInBytes = cuda_pic->w[i] * ((pic->bpc + 7) / 8);
+        m.WidthInBytes = (size_t)cuda_pic->w[i] * ((pic->bpc + 7) / 8);
         m.Height = cuda_pic->h[i];
         if ((bitmask >> i) & 1)
             CHECK_CUDA_RETURN(cu_f, cuMemcpy2DAsync(&m, cuda_priv->cuda.str));
@@ -73,7 +73,7 @@ int vmaf_cuda_picture_upload_async(VmafPicture *cuda_pic, VmafPicture *pic, uint
         m.srcPitch = pic->stride[i];
         m.dstDevice = (CUdeviceptr)cuda_pic->data[i];
         m.dstPitch = cuda_pic->stride[i];
-        m.WidthInBytes = cuda_pic->w[i] * ((pic->bpc + 7) / 8);
+        m.WidthInBytes = (size_t)cuda_pic->w[i] * ((pic->bpc + 7) / 8);
         m.Height = cuda_pic->h[i];
         if ((bitmask >> i) & 1)
             CHECK_CUDA_RETURN(cu_f, cuMemcpy2DAsync(&m, cuda_priv->cuda.str));
@@ -132,8 +132,8 @@ int vmaf_cuda_picture_alloc_pinned(VmafPicture *pic, enum VmafPixelFormat pix_fm
     if (pic->pix_fmt == VMAF_PIX_FMT_YUV400P)
         pic->w[1] = pic->w[2] = pic->h[1] = pic->h[2] = 0;
 
-    const int aligned_y = (pic->w[0] + DATA_ALIGN_PINNED - 1) & ~(DATA_ALIGN_PINNED - 1);
-    const int aligned_c = (pic->w[1] + DATA_ALIGN_PINNED - 1) & ~(DATA_ALIGN_PINNED - 1);
+    const unsigned aligned_y = (pic->w[0] + DATA_ALIGN_PINNED - 1) & ~(DATA_ALIGN_PINNED - 1u);
+    const unsigned aligned_c = (pic->w[1] + DATA_ALIGN_PINNED - 1) & ~(DATA_ALIGN_PINNED - 1u);
     const int hbd = pic->bpc > 8;
     pic->stride[0] = aligned_y << hbd;
     pic->stride[1] = pic->stride[2] = aligned_c << hbd;
@@ -248,7 +248,8 @@ int vmaf_cuda_picture_alloc(VmafPicture *pic, void *cookie)
         }
         CHECK_CUDA_GOTO(cu_f,
                         cuMemAllocPitch((CUdeviceptr *)&pic->data[i], (size_t *)&pic->stride[i],
-                                        pic->w[i] * ((pic->bpc + 7) / 8), pic->h[i], 8 << hbd),
+                                        (size_t)pic->w[i] * ((pic->bpc + 7) / 8), pic->h[i],
+                                        8 << hbd),
                         fail);
     }
 
@@ -273,7 +274,7 @@ int vmaf_cuda_picture_free(VmafPicture *pic, void *cookie)
     if (!pic)
         return -EINVAL;
 
-    int err = vmaf_ref_load(pic->ref);
+    long err = vmaf_ref_load(pic->ref);
     if (!err)
         return -EINVAL;
